Refactor imports in App and Admin components for cleaner structure; update features index to export new components
