+++ src/App.tsx
@@ -1,10 +1,9 @@
-import { useMyState } from "./utils/state.ts"
-
 import { Gallery } from "./pages/Gallery.tsx"
 import { About } from "./pages/About.tsx"
 import { Admin } from "./pages/Admin.tsx"
-import { Header } from "./features/Header.tsx"
-import { Footer } from "./features/Footer.tsx"
+
+import { Header, Footer } from "./features"
+import { useMyState } from "./utils/state.ts"
 import { useKomani } from "./hooks/useKomani.tsx"
 
 function App() {
