+++ src/features/index.tsx
@@ -1,3 +1,8 @@
 import { AdminPhotos } from "./adminPhotos/AdminPhotos"
+import { AdminTags } from "./adminTags/AdminTags"
+import { Banner } from "./Banner"
+import { GalleryCTA } from "./GalleryCTA"
+import { Header } from "./Header"
+import { Footer } from "./Footer"
 
-export { AdminPhotos }
+export { AdminPhotos, AdminTags, Banner, GalleryCTA, Header, Footer }
